refactor(validator/naming): split health host wrapper from pure health logic
