+++ calculogic-validator/bin/calculogic-validator-health.mjs
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
 
-import { runNamingHealthCheckEntrypoint } from '../naming/src/health/naming-health-check.logic.mjs';
+import { runNamingHealthCheckEntrypoint } from '../naming/src/health/naming-health-check.host.mjs';
 
 runNamingHealthCheckEntrypoint();
