+++ calculogic-validator/scripts/validator-health-check.host.mjs
@@ -1,3 +1,3 @@
-import { runNamingHealthCheckEntrypoint } from '../naming/src/health/naming-health-check.logic.mjs';
+import { runNamingHealthCheckEntrypoint } from '../naming/src/health/naming-health-check.host.mjs';
 
 runNamingHealthCheckEntrypoint();
